+++ lib/Analizo.pm
@@ -2,6 +2,7 @@ package Analizo;
 use App::Cmd::Setup -app;
 use strict;
 use warnings;
+use local::lib;
 
 our $VERSION = '1.21.0';